Modfy: KR 関係者削除時にそのユーザーが Objective 責任者となっている下位 Objective が紐付いている場合は削除不可にする

+++ app/controllers/key_results_controller.rb
@@ -104,6 +104,10 @@ class KeyResultsController < ApplicationController
         member.update!(role: role)
       end
     elsif behavior == 'remove'
+      if @key_result.child_objectives.joins(:objective_members).where(objective_members: { user_id: user_id }).exists?
+        @key_result.errors[:error] << '下位 Objective が紐付いているため削除できません'
+        raise
+      end
       # FIXME: 任意のユーザIDで作成してしまうが、サーバ側で採番しない？
       member = @key_result.key_result_members.find_by(user_id: user_id)
       member.destroy!
